+++ src/fr/maxlego08/menu/button/buttons/ZInventoryButton.java
@@ -3,6 +3,7 @@ package fr.maxlego08.menu.button.buttons;
 import fr.maxlego08.menu.api.Inventory;
 import fr.maxlego08.menu.api.InventoryManager;
 import fr.maxlego08.menu.api.button.buttons.InventoryButton;
+import fr.maxlego08.menu.api.command.CommandManager;
 import fr.maxlego08.menu.button.ZButton;
 import fr.maxlego08.menu.inventory.inventories.InventoryDefault;
 import fr.maxlego08.menu.zcore.enums.Message;
@@ -15,24 +16,35 @@ import java.util.Optional;
 public class ZInventoryButton extends ZButton implements InventoryButton {
 
     private final InventoryManager inventoryManager;
+    private final CommandManager commandManager;
     private final String inventoryName;
     private final String pluginName;
+    private final List<String> arguments;
+    private final int toPage;
 
-    /**
-     * @param inventoryManager
-     * @param inventoryName
-     * @param pluginName
-     */
-    public ZInventoryButton(InventoryManager inventoryManager, String inventoryName, String pluginName) {
+    public ZInventoryButton(InventoryManager inventoryManager, CommandManager commandManager, String inventoryName, String pluginName, List<String> arguments, int toPage) {
         super();
         this.inventoryManager = inventoryManager;
+        this.commandManager = commandManager;
         this.inventoryName = inventoryName;
         this.pluginName = pluginName;
+        this.arguments = arguments;
+        this.toPage = toPage;
     }
 
     @Override
     public String getInventory() {
         return this.inventoryName;
+    }
+
+    @Override
+    public List<String> getArguments() {
+        return null;
+    }
+
+    @Override
+    public int getToPage() {
+        return 0;
     }
 
     @Override
@@ -51,8 +63,25 @@ public class ZInventoryButton extends ZButton implements InventoryButton {
             return;
         }
 
+        if (!this.arguments.isEmpty()) {
+
+            for (int i = 0; i < this.arguments.size(); i++) {
+                String name = String.valueOf(i - 4);
+                String argument = this.arguments.get(i);
+
+                if (argument.contains(":")) {
+                    String[] values = argument.split(":", 2);
+                    name = values[0];
+                    argument = values[1];
+                }
+
+                this.commandManager.setPlayerArgument(player, name, argument);
+            }
+
+        }
+
         Inventory toInventory = optional.get();
-        this.inventoryManager.openInventory(player, toInventory, 1, oldInventories);
+        this.inventoryManager.openInventory(player, toInventory, this.toPage, oldInventories);
     }
 
 }